techdebt: [PL-41058]: Package the Entity class leaf nodes of 400-rest
