+++ 400-rest/src/main/java/io/harness/serializer/kryo/ManagerKryoRegistrar.java
@@ -156,7 +156,6 @@ import software.wings.beans.Setup.SetupStatus;
 import software.wings.beans.TechStack;
 import software.wings.beans.TrialSignupOptions;
 import software.wings.beans.User;
-import software.wings.beans.UserGroupEntityReference;
 import software.wings.beans.UserInvite;
 import software.wings.beans.UserInviteSource;
 import software.wings.beans.account.AccountEventType;
@@ -182,6 +181,7 @@ import software.wings.beans.artifact.SftpArtifactStream;
 import software.wings.beans.artifact.SmbArtifactStream;
 import software.wings.beans.command.CommandUnitDetails;
 import software.wings.beans.command.OverridingCommandUnitDescriptor;
+import software.wings.beans.entity.UserGroupEntityReference;
 import software.wings.beans.loginSettings.UserLockoutInfo;
 import software.wings.beans.security.AccountPermissions;
 import software.wings.beans.security.AppPermission;
