+++ 400-rest/src/main/java/io/harness/serializer/morphia/ManagerMorphiaRegistrar.java
@@ -145,7 +145,6 @@ import software.wings.beans.ConfigFile;
 import software.wings.beans.ContainerInfrastructureMapping;
 import software.wings.beans.CustomInfrastructureMapping;
 import software.wings.beans.DelegateSequenceConfig;
-import software.wings.beans.DeletedEntity;
 import software.wings.beans.DeploymentPreference;
 import software.wings.beans.DeploymentSpecification;
 import software.wings.beans.DirectKubernetesInfrastructureMapping;
@@ -265,6 +264,7 @@ import software.wings.beans.container.KubernetesContainerTask;
 import software.wings.beans.container.PcfServiceSpecification;
 import software.wings.beans.container.UserDataSpecification;
 import software.wings.beans.datatretention.LongerDataRetentionState;
+import software.wings.beans.entity.DeletedEntity;
 import software.wings.beans.entityinterface.TagAware;
 import software.wings.beans.governance.GovernanceConfig;
 import software.wings.beans.infrastructure.CloudFormationRollbackConfig;
