--- 400-rest/src/main/java/software/wings/beans/SubEntityType.java
@@ -1,10 +1,0 @@
-/*
- * Copyright 2020 Harness Inc. All rights reserved.
- * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
- * that can be found in the licenses directory at the root of this repository, also available at
- * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
- */
-
-package software.wings.beans;
-
-public enum SubEntityType { NOTIFICATION_GROUP, INFRA_DEFINITION }
