+++ 400-rest/src/main/java/software/wings/beans/entity/BUILD.bazel
@@ -1,0 +1,22 @@
+# Copyright 2023 Harness Inc. All rights reserved.
+# Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
+# that can be found in the licenses directory at the root of this repository, also available at
+# https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
+
+load("@rules_java//java:defs.bzl", "java_library")
+load("//:tools/bazel/macros.bzl", "run_analysis_per_module")
+
+java_library(
+    name = "module",
+    srcs = glob(["*.java"]),
+    visibility = ["//visibility:public"],
+    deps = [
+        "//:lombok",
+        "//960-persistence:module",
+        "//980-commons/src/main/java/io/harness/annotations:module",
+        "//999-annotations/src/main/java/io/harness/annotations/dev:module",
+        "@maven//:dev_morphia_morphia_core",
+    ],
+)
+
+run_analysis_per_module()
