+++ 400-rest/src/main/java/software/wings/beans/entity/SampleAppEntityStatus.java
@@ -1,25 +1,21 @@
 /*
- * Copyright 2022 Harness Inc. All rights reserved.
+ * Copyright 2020 Harness Inc. All rights reserved.
  * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
  * that can be found in the licenses directory at the root of this repository, also available at
  * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
  */
 
-package software.wings.beans;
-
-import static io.harness.annotations.dev.HarnessTeam.PL;
-
-import io.harness.annotations.dev.OwnedBy;
+package software.wings.beans.entity;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
-@OwnedBy(PL)
-public class UserGroupEntityReference {
-  private String id;
-  private String appId;
-  private String accountId;
+public class SampleAppEntityStatus {
+  private String entityName;
   private String entityType;
+  private Health health;
+  public enum Health { GOOD, BAD }
+  ;
 }
