+++ 400-rest/src/main/java/software/wings/beans/entity/SubEntityType.java
@@ -5,20 +5,6 @@
  * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
  */
 
-package software.wings.beans;
+package software.wings.beans.entity;
 
-import lombok.Builder;
-import lombok.Data;
-
-/**
- * Generic entity reference
- * @author rktummala on 12/08/17
- */
-@Data
-@Builder
-public class EntityReference {
-  private String id;
-  private String name;
-  private String appId;
-  private String entityType;
-}
+public enum SubEntityType { NOTIFICATION_GROUP, INFRA_DEFINITION }
