+++ 400-rest/src/main/java/software/wings/beans/entity/UserGroupEntityReference.java
@@ -1,31 +1,25 @@
 /*
- * Copyright 2020 Harness Inc. All rights reserved.
+ * Copyright 2022 Harness Inc. All rights reserved.
  * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
  * that can be found in the licenses directory at the root of this repository, also available at
  * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
  */
 
-package software.wings.beans.security.restrictions;
+package software.wings.beans.entity;
 
 import static io.harness.annotations.dev.HarnessTeam.PL;
 
 import io.harness.annotations.dev.OwnedBy;
 
-import software.wings.beans.entity.EntityReference;
-
-import java.util.Set;
 import lombok.Builder;
 import lombok.Data;
 
-/**
- * @author rktummala on 06/24/18
- */
-@OwnedBy(PL)
 @Data
 @Builder
-public class AppRestrictionsSummary {
-  private EntityReference application;
-  private boolean hasAllProdEnvAccess;
-  private boolean hasAllNonProdEnvAccess;
-  private Set<EntityReference> environments;
+@OwnedBy(PL)
+public class UserGroupEntityReference {
+  private String id;
+  private String appId;
+  private String accountId;
+  private String entityType;
 }
