+++ 400-rest/src/main/java/software/wings/beans/security/UserGroup.java
@@ -33,7 +33,7 @@ import software.wings.beans.Base;
 import software.wings.beans.NotificationChannelType;
 import software.wings.beans.NotificationReceiverInfo;
 import software.wings.beans.User;
-import software.wings.beans.UserGroupEntityReference;
+import software.wings.beans.entity.UserGroupEntityReference;
 import software.wings.beans.notification.NotificationSettings;
 import software.wings.beans.notification.SlackNotificationSetting;
 import software.wings.beans.sso.SSOType;
