+++ 400-rest/src/main/java/software/wings/helpers/ext/account/DeleteAccountHelper.java
@@ -36,10 +36,10 @@ import io.harness.scheduler.PersistentScheduler;
 import software.wings.beans.Account;
 import software.wings.beans.Application;
 import software.wings.beans.Application.ApplicationKeys;
-import software.wings.beans.DeletedEntity;
-import software.wings.beans.DeletedEntity.DeletedEntityKeys;
-import software.wings.beans.DeletedEntity.DeletedEntityType;
 import software.wings.beans.User;
+import software.wings.beans.entity.DeletedEntity;
+import software.wings.beans.entity.DeletedEntity.DeletedEntityKeys;
+import software.wings.beans.entity.DeletedEntity.DeletedEntityType;
 import software.wings.beans.entityinterface.ApplicationAccess;
 import software.wings.beans.sso.SSOSettings;
 import software.wings.scheduler.events.segment.SegmentGroupEventJobContext;
