+++ 400-rest/src/main/java/software/wings/service/impl/HarnessSampleAppServiceImpl.java
@@ -31,14 +31,14 @@ import software.wings.beans.Application;
 import software.wings.beans.EntityType;
 import software.wings.beans.Environment;
 import software.wings.beans.Pipeline;
-import software.wings.beans.SampleAppEntityStatus;
-import software.wings.beans.SampleAppEntityStatus.Health;
 import software.wings.beans.SampleAppStatus;
 import software.wings.beans.Service;
 import software.wings.beans.SettingAttribute;
 import software.wings.beans.SettingAttribute.SettingCategory;
 import software.wings.beans.Workflow;
 import software.wings.beans.artifact.ArtifactStream;
+import software.wings.beans.entity.SampleAppEntityStatus;
+import software.wings.beans.entity.SampleAppEntityStatus.Health;
 import software.wings.infra.InfrastructureDefinition;
 import software.wings.service.intfc.AccountService;
 import software.wings.service.intfc.AppService;
