+++ 400-rest/src/main/java/software/wings/service/impl/UserGroupServiceImpl.java
@@ -80,8 +80,8 @@ import software.wings.beans.EntityType;
 import software.wings.beans.Event.Type;
 import software.wings.beans.User;
 import software.wings.beans.User.UserKeys;
-import software.wings.beans.UserGroupEntityReference;
 import software.wings.beans.UserInvite;
+import software.wings.beans.entity.UserGroupEntityReference;
 import software.wings.beans.notification.NotificationSettings;
 import software.wings.beans.security.AccountPermissions;
 import software.wings.beans.security.AppPermission;
