+++ 400-rest/src/test/java/software/wings/helpers/ext/account/DeleteAccountHelperTest.java
@@ -48,11 +48,11 @@ import software.wings.WingsBaseTest;
 import software.wings.beans.Account;
 import software.wings.beans.AccountType;
 import software.wings.beans.Application;
-import software.wings.beans.DeletedEntity;
 import software.wings.beans.LicenseInfo;
 import software.wings.beans.ServiceTemplate;
 import software.wings.beans.User;
 import software.wings.beans.account.AccountStatus;
+import software.wings.beans.entity.DeletedEntity;
 import software.wings.beans.entityinterface.ApplicationAccess;
 import software.wings.licensing.LicenseService;
 import software.wings.scheduler.events.segment.SegmentGroupEventJobContext;
