+++ 400-rest/src/test/java/software/wings/service/impl/HarnessSampeAppServiceTest.java
@@ -46,13 +46,13 @@ import software.wings.beans.EntityType;
 import software.wings.beans.Environment;
 import software.wings.beans.Environment.Builder;
 import software.wings.beans.Pipeline;
-import software.wings.beans.SampleAppEntityStatus;
-import software.wings.beans.SampleAppEntityStatus.Health;
 import software.wings.beans.SampleAppStatus;
 import software.wings.beans.Service;
 import software.wings.beans.SettingAttribute;
 import software.wings.beans.SettingAttribute.SettingCategory;
 import software.wings.beans.Workflow;
+import software.wings.beans.entity.SampleAppEntityStatus;
+import software.wings.beans.entity.SampleAppEntityStatus.Health;
 import software.wings.dl.WingsPersistence;
 import software.wings.infra.InfrastructureDefinition;
 import software.wings.service.intfc.AppService;
