+++ 400-rest/src/test/java/software/wings/service/impl/UserGroupServiceImplTest.java
@@ -116,7 +116,7 @@ import software.wings.beans.Event.Type;
 import software.wings.beans.LicenseInfo;
 import software.wings.beans.Role;
 import software.wings.beans.User;
-import software.wings.beans.UserGroupEntityReference;
+import software.wings.beans.entity.UserGroupEntityReference;
 import software.wings.beans.notification.NotificationSettings;
 import software.wings.beans.notification.SlackNotificationSetting;
 import software.wings.beans.security.AccountPermissions;
