+++ 400-rest/src/test/java/software/wings/service/impl/expression/ExpressionBuilderServiceTest.java
@@ -84,10 +84,10 @@ import software.wings.beans.PipelineStage.PipelineStageElement;
 import software.wings.beans.Service;
 import software.wings.beans.ServiceTemplate;
 import software.wings.beans.ServiceVariable;
-import software.wings.beans.SubEntityType;
 import software.wings.beans.Variable;
 import software.wings.beans.VariableType;
 import software.wings.beans.Workflow;
+import software.wings.beans.entity.SubEntityType;
 import software.wings.service.intfc.AccountService;
 import software.wings.service.intfc.AppService;
 import software.wings.service.intfc.PipelineService;
